moved check_block_size
function moved to utils unit
