+++ pyssage/utils.py
@@ -142,3 +142,23 @@ def create_output_table(output_text: list, table_data: list, col_headers: list, 
     # create table data
     for row in table_data:
         output_text.append(template.format(*row))
+
+
+def check_block_size(max_block_size: int, n: int, x: int) -> int:
+    """
+    Check the maximum block size to be sure it doesn't exceed limits for the particular analysis and input data
+
+    :param max_block_size: the requested largest block size
+    :param n: the length of the transect
+    :param x: the number of "blocks" that make up the analysis; this affects the maximum allowable size
+    :return: the maximum block size that will actually be used in the analysis
+    """
+    if max_block_size == 0:
+        max_block_size = n // x
+    if max_block_size < 2:
+        max_block_size = 2
+    elif max_block_size > n // x:
+        max_block_size = n // x
+        print("Maximum block size cannot exceed {:0.1f}% of transect length. Reduced to {}.".format(100 / x,
+                                                                                                    max_block_size))
+    return max_block_size
